+++ src/main/java/plugily/projects/thebridge/events/Events.java
@@ -41,10 +41,10 @@ import org.bukkit.event.player.PlayerDropItemEvent;
 import org.bukkit.event.player.PlayerInteractEvent;
 import org.bukkit.event.player.PlayerTeleportEvent;
 import org.bukkit.inventory.ItemStack;
-import pl.plajerlair.commonsbox.minecraft.compat.VersionUtils;
-import pl.plajerlair.commonsbox.minecraft.compat.events.api.CBPlayerInteractEvent;
-import pl.plajerlair.commonsbox.minecraft.compat.events.api.CBPlayerSwapHandItemsEvent;
-import pl.plajerlair.commonsbox.minecraft.compat.xseries.XMaterial;
+import plugily.projects.commonsbox.minecraft.compat.VersionUtils;
+import plugily.projects.commonsbox.minecraft.compat.events.api.CBPlayerInteractEvent;
+import plugily.projects.commonsbox.minecraft.compat.events.api.CBPlayerSwapHandItemsEvent;
+import plugily.projects.commonsbox.minecraft.compat.xseries.XMaterial;
 import plugily.projects.thebridge.ConfigPreferences;
 import plugily.projects.thebridge.Main;
 import plugily.projects.thebridge.arena.Arena;
@@ -53,6 +53,7 @@ import plugily.projects.thebridge.arena.ArenaRegistry;
 import plugily.projects.thebridge.arena.ArenaState;
 import plugily.projects.thebridge.arena.ArenaUtils;
 import plugily.projects.thebridge.handlers.items.SpecialItemManager;
+import plugily.projects.thebridge.utils.Debugger;
 import plugily.projects.thebridge.utils.Utils;
 
 /**
@@ -173,12 +174,16 @@ public class Events implements Listener {
       event.setCancelled(true);
       event.setFoodLevel(20);
     }
-    if(VersionUtils.getItemInHand(player).getType() == XMaterial.GOLDEN_APPLE.parseMaterial()) {
-      event.setFoodLevel(20);
-    }
     if(arena.getArenaState() != ArenaState.IN_GAME) {
       event.setFoodLevel(20);
       event.setCancelled(true);
+    }
+    if(event.getItem() == null) {
+      return;
+    }
+    if(event.getItem().getType() == XMaterial.GOLDEN_APPLE.parseMaterial()) {
+      event.setFoodLevel(20);
+      player.setHealth(VersionUtils.getMaxHealth(player));
     }
   }
 
@@ -260,6 +265,9 @@ public class Events implements Listener {
       return;
     }
     if(event.getCause().equals(PlayerTeleportEvent.TeleportCause.END_PORTAL)) {
+      event.setCancelled(true);
+    }
+    if(event.getCause().equals(PlayerTeleportEvent.TeleportCause.NETHER_PORTAL)) {
       event.setCancelled(true);
     }
   }